Add errors to placing trade
